Introduce an abstraction around big integers
I am "shopping around" for a new `BigInteger` implementation, and one of the ways to make that simpler is to abstract away that type.

The implementation classes are hopefully-inlined value classes, so the performance overhead should be minimal, and the benefit is we get a cleaner API when calling our `BigInt` interface, as we can move our extension functions and other hacks all into the one interface.